perf: Load bookings used for limits in bulk
